avanzamenti buy page, struttura generale db/relazioni jakarta

+++ ddl/dbproj/optservice.sql
@@ -1,9 +1,11 @@
+use dbproj;
+
 create table optservice
 (
     idOptService int auto_increment
         primary key,
     name         varchar(45) not null,
-    monthly      int         not null,
+    monthly      float         not null,
     refEmployee  int         null,
     constraint Name_UNIQUE
         unique (name),
